Pull the exhaustive test into a separate file
